setup gpu

+++ src/nn/msg_pass.cpp
@@ -43,7 +43,7 @@ void Node2NodeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 	
 	int nnz = 0;
 	auto& data = this->cpu_weight->data;
-	for (int i = 0; i < graph->num_nodes; ++i)
+	for (uint i = 0; i < graph->num_nodes; ++i)
 	{
 		data->row_ptr[i] = nnz;
 		auto& list = graph->in_edges->head[i];
@@ -54,7 +54,7 @@ void Node2NodeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 			nnz++;
 		}
 	}
-	assert(nnz == graph->num_edges);
+	assert(nnz == (int)graph->num_edges);
 	data->row_ptr[graph->num_nodes] = nnz;
 }
 
@@ -70,7 +70,7 @@ void Edge2NodeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 	
 	int nnz = 0;
 	auto& data = this->cpu_weight->data;		
-	for (int i = 0; i < graph->num_nodes; ++i)
+	for (uint i = 0; i < graph->num_nodes; ++i)
 	{
 		data->row_ptr[i] = nnz;
 		auto& list = graph->in_edges->head[i];
@@ -81,7 +81,7 @@ void Edge2NodeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 			nnz++;
 		}
 	}
-	assert(nnz == graph->num_edges);
+	assert(nnz == (int)graph->num_edges);
 	data->row_ptr[graph->num_nodes] = nnz;    
 }
 
@@ -97,7 +97,7 @@ void Node2EdgeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
     this->cpu_weight->ResizeSp(graph->num_edges, graph->num_edges + 1);
         
 	auto& data = this->cpu_weight->data;
-    for (int i = 0; i < graph->num_edges; ++i)
+    for (uint i = 0; i < graph->num_edges; ++i)
     {
         data->row_ptr[i] = nnz;
         data->val[nnz] = 1.0;
@@ -118,7 +118,7 @@ void Edge2EdgeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
     int nnz = 0;
     this->cpu_weight->Reshape({graph->num_edges, graph->num_edges});
     size_t cnt = 0;
-    for (int i = 0; i < graph->num_nodes; ++i)
+    for (uint i = 0; i < graph->num_nodes; ++i)
     {
         auto in_cnt = graph->in_edges->head[i].size();
         cnt += in_cnt * (in_cnt - 1); 
@@ -126,7 +126,7 @@ void Edge2EdgeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
     this->cpu_weight->ResizeSp(cnt, graph->num_edges + 1);            
         
     auto& data = this->cpu_weight->data;
-    for (int i = 0; i < graph->num_edges; ++i)
+    for (uint i = 0; i < graph->num_edges; ++i)
     {
         data->row_ptr[i] = nnz;
         int node_from = graph->edge_list[i].first, node_to = graph->edge_list[i].second; 
@@ -142,7 +142,7 @@ void Edge2EdgeMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
     }
     data->row_ptr[graph->num_edges] = nnz;
     assert(nnz == data->nnz);
-    assert(data->nnz == cnt);
+    assert(data->nnz == (int)cnt);
 }
 
 template class Edge2EdgeMsgPass<CPU, double>;
@@ -157,7 +157,7 @@ void SubgraphMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 	
 	int nnz = 0;
 	auto& data = this->cpu_weight->data;
-	for (int i = 0; i < graph->num_subgraph; ++i)
+	for (uint i = 0; i < graph->num_subgraph; ++i)
 	{
 		data->row_ptr[i] = nnz;
 		auto& list = graph->subgraph->head[i];
@@ -168,7 +168,7 @@ void SubgraphMsgPass<mode, Dtype>::InitCPUWeight(GraphStruct* graph)
 			nnz++;
 		}
 	}
-	assert(nnz == graph->num_nodes);
+	assert(nnz == (int)graph->num_nodes);
 	data->row_ptr[graph->num_subgraph] = nnz;
 }
